2010-06-20 17:41 UTC+0200 Viktor Szakats (harbour.01 syenar.hu)   * src/pp/hbpp.c   * src/vm/cmdarg.c   * src/rtl/defpath.c   * src/rtl/hbstrfmt.c   * src/rtl/dates.c   * src/rtl/hbinet.c   * src/rtl/gtpca/gtpca.c   * src/rtl/file.c   * src/compiler/ppcomp.c      + Added strlen() casts to pacify warnings
  * src/vm/hashfunc.c
    + Using hb_vmPushSize() to add HB_SIZE support for Win64

  * src/vm/debug.c
    ! HB_SIZE -> HB_LONG. Signaled by warning

  * include/hbvm.h
  * src/vm/hvm.c
    + Added hb_vmPushSize()
      Please review, I copied hb_vmPushHBLong()

  * src/common/hbfopen.c
    + Suppressed MS warning about "unsafe" fopen()
      Also readded fopen_s() code, which is made inactive permanently.

  * include/hbexprop.h
    + Added argument names to function declarations.

  * contrib/hbct/ctwin.c
    + HB_SIZE -> HB_UINT, plus added necessary casts. Signaled by warning.

  * contrib/xhb/xhbarr.c
    + HB_SIZE -> int, plus added necessary casts. Signaled by warning.

  * contrib/hbsqlit3/hbsqlit3.c
    - Suppressed MS warning about "unsafe" fopen()

+++ harbour/src/compiler/ppcomp.c
@@ -124,7 +124,7 @@ static HB_BOOL hb_pp_CompilerSwitch( void * cargo, const char * szSwitch,
 
    iValue = *piValue;
 
-   i = strlen( szSwitch );
+   i = ( int ) strlen( szSwitch );
    if( i > 1 && ( ( int ) ( szSwitch[ i - 1 ] - '0' ) ) == iValue )
       --i;
