+++ harbour/src/rtl/dates.c
@@ -80,7 +80,7 @@ char * hb_dateFormat( const char * szDate, char * szFormattedDate, const char * 
    /*
     * Determine the maximum size of the formatted date string
     */
-   size = strlen( szDateFormat );
+   size = ( int ) strlen( szDateFormat );
    if( size > 10 )
       size = 10;
 
@@ -259,7 +259,7 @@ int hb_dateUnformatRaw( const char * szDate, const char * szDateFormat, long * p
 
       if( ! szDateFormat )
          szDateFormat = hb_setGetDateFormat();
-      size = strlen( szDateFormat );
+      size = ( int ) strlen( szDateFormat );
 
       for( count = used = 0; count < size && used < 3; count++ )
       {
@@ -309,7 +309,7 @@ int hb_dateUnformatRaw( const char * szDate, const char * szDateFormat, long * p
       /* If there are non-digits at the start of the date field,
          they are not to be treated as date field separators */
       non_digit = 1;
-      size = strlen( szDate );
+      size = ( int ) strlen( szDate );
       for( count = used = 0; count < size; count++ )
       {
          digit = szDate[ count ];
@@ -394,7 +394,7 @@ char * hb_timeFormat( char * szBuffer, const char * szTimeFormat, long lMilliSec
    hb_timeDecode( lMilliSec, &iHour, &iMinutes, &iSeconds, &iMSec );
    szTimeBuffer = szBuffer;
 
-   size = hb_strnlen( szTimeFormat, 16 );
+   size = ( int ) hb_strnlen( szTimeFormat, 16 );
    iPM = i12 = 0;
    for( i = 0; i < size; ++i )
    {
@@ -538,7 +538,7 @@ long hb_timeUnformat( const char * szTime, const char * szTimeFormat )
    if( ! szTimeFormat )
       szTimeFormat = hb_setGetTimeFormat();
 
-   size = hb_strnlen( szTime, hb_strnlen( szTimeFormat, 16 ) );
+   size = ( int ) hb_strnlen( szTime, hb_strnlen( szTimeFormat, 16 ) );
    iHour = iMinutes = iSeconds = iMSec = iPM = -1;
    prec = 0;
    for( i = count = 0; i < size && szTime[ count ]; ++i )
