+++ harbour/src/rtl/gtpca/gtpca.c
@@ -330,7 +330,7 @@ static void hb_gt_pca_AnsiSetCursorPos( int iRow, int iCol )
    {
       char buff[ 16 ];
       hb_snprintf( buff, sizeof( buff ), "\x1B[%d;%dH", iRow + 1, iCol + 1 );
-      hb_gt_pca_termOut( buff, strlen( buff ) );
+      hb_gt_pca_termOut( buff, ( int ) strlen( buff ) );
       s_iRow = iRow;
       s_iCol = iCol;
    }
